refactor: rename navigation components for consistent naming
