+++ src/components/navigation/SectionNavButton.tsx
@@ -1,4 +1,4 @@
-// src/components/ui/SectionNavButton.tsx
+// src/components/navigation/SectionNavButton.tsx
 
 import Link from 'next/link';
 import type { SectionNavButtonProps } from '~/types/navigation';
